requirements.txt
Flask==0.10.1
Pillow==3.1.1